feat(profile-card): migrates to material ui
Migrates the ProfileCard to material ui

re #144
